requirements.txt
streamlit
psutil
pynvml
pandas
altair
streamlit-autorefresh
mysql-connector-python
